+++ src/main/java/icons/KdbIcons.java
@@ -75,6 +75,7 @@ public final class KdbIcons {
 
         public static final Icon OpenInEditor = AllIcons.General.FitContent;
         public static final Icon FlipTable = load("/org/kdb/inside/brains/icons/flipTable.svg");
+        public static final Icon TableIndex = load("/org/kdb/inside/brains/icons/tableIndex.svg");
     }
 
     public static final class Node {